Function to calculate linear distance between bus and stop

+++ src/PassengerAgent.java
@@ -112,7 +112,7 @@ public class PassengerAgent extends Agent {
      */
     private class RequestPerformer extends Behaviour {
         private AID bestBus; // The agent who provides the best offer
-        private int bestTime;  // The best offered time
+        private double bestTime;  // The best offered time
         private int repliesCnt = 0; // The counter of replies from bus agents
         private MessageTemplate mt; // The template to receive replies
         private int step = 0;
@@ -143,7 +143,7 @@ public class PassengerAgent extends Agent {
                         // Reply received
                         if (reply.getPerformative() == ACLMessage.PROPOSE) {
                             // This is an offer
-                            int time = Integer.parseInt(reply.getContent());
+                            double time = Double.parseDouble(reply.getContent());
                             if (bestBus == null || time < bestTime) {
                                 // This is the best offer at present
                                 bestTime = time;
